Improve Bootstrap JavaScript Components
Moved the table reservation form from
the last content row into a modal.

Included a radio button group in the table reservation form
to enable diners to ask for a table
in the smoking/non-smoking section of the restaurant.

Removed the tooltip from the Reserve Table button.
Updated the Reserve Table button to show the modal
containing the table reservation form when the button is clicked.